Modified for MSVC.

+++ Source/Core/FileFormat/KVSML/ColorMapTag.cpp
@@ -187,7 +187,7 @@ const bool ColorMapValueTag::read( const kvs::XMLElement::SuperClass* element )
 
     // s ="xxx"
     const std::string s = kvs::XMLElement::AttributeValue( element, "s" );
-    if ( s != "" ) m_scalar = atof( s.c_str() );
+    if ( s != "" ) m_scalar = static_cast<float>( atof( s.c_str() ) );
     else
     {
         kvsMessageError( "'s' is not specified in <%s>.", tag_name.c_str() );
